Prune unused actor sampling annotations
