+++ fdbclient/ActorLineageProfiler.cpp
@@ -20,7 +20,6 @@
 
 #include "flow/flow.h"
 #include "flow/singleton.h"
-#include "flow/IAsyncFile.h"
 #include "fdbclient/ActorLineageProfiler.h"
 #include "NameLineage.h"
 #include <msgpack.hpp>
@@ -291,11 +290,6 @@ struct ProfilerImpl {
 };
 
 ActorLineageProfilerT::ActorLineageProfilerT() : impl(new ProfilerImpl()) {
-	// collection->collector()->addGetter(WaitState::Network,
-	//                                    std::bind(&ActorLineageSet::copy, std::ref(g_network->getActorLineageSet())));
-	// collection->collector()->addGetter(
-	//     WaitState::Disk,
-	//     std::bind(&ActorLineageSet::copy, std::ref(IAsyncFileSystem::filesystem()->getActorLineageSet())));
 	collection->collector()->addGetter(WaitState::Running, []() {
 		return std::vector<Reference<ActorLineage>>({ SampleCollection::instance().getLineage() });
 	});
