+++ fdbclient/include/fdbclient/ProcessInterface.h
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-#include "fdbclient/AnnotateActor.h"
+#include "fdbclient/WaitState.h"
 #include "fdbclient/FDBTypes.h"
 #include "fdbrpc/fdbrpc.h"
 #include "fdbclient/WellKnownEndpoints.h"
